Move default actions
